record done=True for all steps until reset happens; algos comput valid using done from replay; no more store valid in replay

+++ rlpyt/algos/utils.py
@@ -49,6 +49,6 @@ def discount_return_n_step(reward, done, n_step, discount, return_dest=None,
 
 
 def valid_from_done(done):
-    valid = torch.ones_like(done)
+    valid = torch.ones_like(done, dtype=torch.float)
     valid[1:] = 1 - torch.clamp(torch.cumsum(done[:-1], dim=0), max=1)
     return valid
